Reapply "[CXP-2599] Enable WLM for process collection by default on linux" (#40643)
The initial [PR](https://github.com/DataDog/datadog-agent/pull/40403)
was reverted due to introducing flakes on the
`TestLinuxTestSuite/TestManualProcessCheckWithIO` e2e test. Since the
change resulted in the process-agent using remote WLM for process data,
there is a race condition between the run of the check command on the
process agent and whether the collectors actually ran on the core agent.
The `workloadmeta.IsInitialized()` will only check if the collectors
were started in the current runtime not across others, so we can't rely
on it. To fix this, we introduce retries in the test since the race
condition has more impact on agent startup.

Fix test flakes + original motivations from
https://github.com/DataDog/datadog-agent/pull/40403
e2e tests + unit tests

e2e tests were run multiple times in the hopes of catching flakes. The
problematic tests are also marked as flaky to prevent issues on main
while we determine if the fix works.

(cherry picked from commit 097053a63b479dfbbdb03464061b523c811d1868)