dyninst/rcscrape: mark if a data item had a failed read
We've seen data items missing related to a runtime ID. Perhaps we
failed to read those memory addresses. We'd like to know about that.

Note that this same treatment should get applied to regular decoding
(perhaps in a more principled way). That's left for a different change
that won't be backported.