dyninst/rcscrape: gracefully handle decoding failures
If we can't decode a message, we don't want to shut down the entire
dyninst subsystem. The fact that that's what happens when decoding
fails is not great, but it's not for this change we intend to backport.

There's an upcoming refactor to address that.

Fixes [DEBUG-4455](https://datadoghq.atlassian.net/browse/DEBUG-4455).