Add `image_resolution_attempts` telemetry for `ImageResolver.Resolve()` (#40697)
### What does this PR do?
Adds `image_resolution_attempts` telemetry to the newly added
`ImageResolver.Resolve()` function to track image resolution attempts,
and whether they resulted in a successful image digest resolution, or if
they defaulted to using the mutable tag.

### Motivation
This telemetry is necessary for the new K8s SSI gradual rollout feature
in order to determine whether a rollout is working successfully or not.
The goal is for this to be easier to display on a dashboard:

| `registry` | `repository` | `digest_resolution` | `outcome` | Notes |
|----------|----------|----------|----------|----------|
| `gcr.io/datadoghq` | `dd-lib-python-init` | `enabled` |
`sha256:abc123` | ✅ Was supposed to resolve, did resolve |
| `hub.docker.com/r/datadog` | `dd-lib-java-init` | `enabled` | `v2` | ❌
Was supposed to resolve, but it did NOT resolve |
| `mycustomregistry.org` | `dd-lib-php-init` | `enabled` | `v2` | ✅
Cannot resolve for custom registry, did NOT resolve |
| `gallery.ecr.aws/datadog` | `dd-lib-rb-init` | `disabled` | `v1` | ✅
Was NOT supposed to resolve, did NOT resolve |
| `gcr.io/datadoghq` | `dd-lib-dotnet-init` | `disabled` |
`sha256:abc123` | ❌ Was NOT supposed to resolve, did resolve *this
shouldn't be possible |


### Describe how you validated your changes
Ran a local app via `injector-dev` to verify that the telemetry counts
for `apm-inject` and `dd-lib-python-init` were accurate.

### Additional Notes