dyninst/object: do not munmap nil (#40780)
Before this change, if an error occurred during decompression but before
we had mmapped the file, we would end up hitting this code path and
logging about it. We don't need to do that.

We were hitting the error because of
https://github.com/DataDog/datadog-agent/pull/40765, so I don't expect
to need to backport this (no harm was caused by it), but still worth
merging.