dyninst/rcscrape: gracefully handle decoding failures (#40779)
#### dyninst/rcscrape: mark if a data item had a failed read

We've seen data items missing related to a runtime ID. Perhaps we
failed to read those memory addresses. We'd like to know about that.

Note that this same treatment should get applied to regular decoding
(perhaps in a more principled way). That's left for a different change
that won't be backported.

#### dyninst/rcscrape: gracefully handle decoding failures

If we can't decode a message, we don't want to shut down the entire
dyninst subsystem. The fact that that's what happens when decoding
fails is not great, but it's not for this change we intend to backport.

There's an upcoming refactor to address that.

Fixes [DEBUG-4455](https://datadoghq.atlassian.net/browse/DEBUG-4455).

[DEBUG-4455]:
https://datadoghq.atlassian.net/browse/DEBUG-4455?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

(cherry picked from commit d4093bae378fe711d468be51d257e2ed1d9c5eb7)