fakeintake client parse host-tags payloads (#40455)
### What does this PR do?

Add necessary code in the `fakeintake` library to parse and deserialize
the host-impl tags.

This requires some filtering, the route to read the tags contains other
payloads that could match while calling `Unmarshall`.

The field `meta` is located in the payload we try to extract, but it's
also located in some other struct that we don't want.

Use an inner filed allocated with a pointer to check for its presence
and therefore validate if we are looking at the right payload.

Add necessary code to query the host-implementation tags from the
command line.

Once we have a match return all payloads.

### Motivation

This allows us later to add some end2end tests for host-tags

### Describe how you validated your changes

1. deploy a test stack using `pulumi`
2. use the `fakeintakectl` to query the host tags
3. see the host-tags in the output

command:
```
./test/fakeintake/build/fakeintakectl --url "http://w.x.y.z:80" get host-infos
```

output:
```
[
  {
    "host-tags": {
      "system": [
        "arch:amd64",
        "cluster_name:gcp-gcp-alexandre-lavigne-gke-deb76db",
        "kube_cluster_name:gcp-gcp-alexandre-lavigne-gke-deb76db",
        "kube_node:gke-gcp-gcp-alexandre-la-default-pool-7a56dc29-8ggv",
        "orch_cluster_id:f340972a-1311-47b3-a6af-c6f83a7ece1c",
        "os:linux"
      ]
    },
    "internalHostname": "gke-gcp-gcp-alexandre-la-default-pool-7a56dc29-8ggv.c.datadog-agent-sandbox.internal"
  }
]
```

### Additional Notes

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>