[CWS-5081] Fix wrong kill status when process failed to be killed (#40632)
### What does this PR do?
**This PR fixes a bug.**
During a kill action, if the kill failed, the action was still flagged
as "performed".

**Details of the bug fix:**

1) When a process needs to be killed but the kill action must be delayed
(by the disarmer), it is placed in the KillQueue.

2) When the process is allowed to be killed, the kill is attempted on
all processes in the queue.

3) The algorithm then assumed that every process in this map was
successfully killed (and flagged the actions as "performed"), without
checking the return value of the actual kill.

This PR introduces a mechanism to check whether the kill was successful.
If the kill fails, the process is removed from the queue and its status
is properly updated to "error".
### Motivation

### Describe how you validated your changes

### Additional Notes