[RC] Store update config version even if config contents don't change
Previously, we would only update information about a config in the agent
client if the config contents didn't change. However, there are
operations against the remote config backend which can still cause the
config version to increase without changing the config contents. Product
teams often want to check if the latest config has been sent everywhere
by querying the state of what RC clients report. With the previous
behavior, updates not touching the config body would cause it to seem
like clients were on an older version, and force product teams to
determine if the config contents at that version is equal to whatever
the current latest version is.

This change causes the RC client in the agent to update the version if
the body dooesn't change. This does not invoke any business logic
callbacks, as nothing has changed in the config contents to warrant
changing configs. It's solely for a better UX for product team UI's.