CONTINT-4836 - fakeintake retunr host-tags for given hostname
Add new parameter `hostname` to fakeintake API `GetHostTags`.

This allows the caller to only return the host-tags for that given
hostname if it exists.
Otherwise return an empty list.

Add a method to return all known hostnames from the list of host-tags we
have in the store.

extract the OS info from the payload, that will be useful for end2end
tests to check which tag set we expect.

Renamed all functions/method to reflect that it only return host-tags
and *not* the whole host infos.

This will allow us to improve our test coverage by testing host-tags
based on the host we query.

Closes: CONTINT-4836

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>