disable preaggregation endpoint for otel-agent (#40847)
### What does this PR do?

Disables preaggregation pipeline feature in the otel agent.

### Motivation

The feature needs an API key of its own, but the otel agent does not
decrypt secrets from the agent config file on its own. The workaround of
loading the new secret from the core agent config was deemed too
invasive and not worthwhile for an experimental feature.

### Describe how you validated your changes

Running the otel agent locally with a core agent config enabling
preaggregation, ensuring that the otel agent did not pick up the
preaggregation endpoint or attempt to send any data.

### Additional Notes