[CWS] Try to fix TestFilterOpenLeafDiscarderActivityDump flackyness (#40635)
### What does this PR do?

It does fix the test flackyness by calling the lost event traced cgroup
cleanup func (`SnapshotTracedCgroups`) when we call
`StartADockerGetDump`. The issue was that, at module startup the traced
cgroup maps was filled with running cgroups, then when the test started
trying to start a docker and get it traced it happen often that the map
was already full and not cleaned up yet to evict cgroups to let the
containers been traced.
At the end it was a test issue, as on a living system-probe the cleanup
callback will eventually be called to evict unwanted entries in the map.

PLUS:
* Fix some trace cgroup map poisoning on CentOS7 where pid 1 has no
cgroup.
* Fix some data races around the cgroup resolver saving/sharing
pointers, now it stores/gives a copy.
* Fix the cleanup of lost cgroup tracing events where, if any, as the
containerID was not filled .. it evicted running container from the
traced cgroup map.

### Motivation

Fix test `TestFilterOpenLeafDiscarderActivityDump`  flackyness

### Describe how you validated your changes

### Additional Notes