Update `ImageResolver` cache structure to use repo name as key (#40762)
### What does this PR do?
Restructures the `ImageResolver` cache structure so that the keys are
only the repository names (ex - `dd-lib-python-init`) and not the
repository URL (ex - `gcr.io/dd-lib-python-init`).

It also enables us to configure the default Datadog container registries
via configuration. This is not public-facing, and will be used primarily
for dogfooding the new feature on staging (given staging uses different
container registries).

### Motivation
The repository URL field in the `K8S_INJECTION_DD` remote config data
was not designed to consumed by DCA, but primarily for use by our
internal `equilibrium` tooling. This meant that by using the repository
URL for the key, it would be limited to only allowing gradual rollout
for customers using `grc.io/datadoghq` as their registry. See
[here](https://docs.datadoghq.com/tracing/trace_collection/automatic_instrumentation/single-step-apm/kubernetes/?tab=agentv764recommended#change-the-default-image-registry)
for other valid Datadog registries.

### Describe how you validated your changes
- Updated existing tests
- Added new unit tests
- Local E2E testing with `injector-dev`

### Additional Notes