[CWS] fix comm confusion between TestPrCtl and TestSnapshot (#40857)
### What does this PR do?

TestPrCtl is using PR_SET_NAME on the testsuite itself, this is causing
issue because a lot of other tests are actually looking for a comm of
the testsuite being "testsuite". To fix this, this PR moves the
PR_SET_NAME to the syscall tester, ensuring this modification is scoped
to this test, and this test only.

### Motivation

### Describe how you validated your changes

### Additional Notes