Add package interaction metadata and config file to ddot-byoc builds (#40805)
### What does this PR do?

- Adds a depends relationship from the ddot-byoc to the datadog-agent in
a way similar to what our official ddot package does.
- Adds a conflict from ddot-byoc towards our official ddot package to
have package managers generally prevent having the default and the
custom ddot installed on the same system.
- It adds the config file that's generated by the build command to the
package.

### Motivation

[ABLD-102](https://datadoghq.atlassian.net/browse/ABLD-102)

### Describe how you validated your changes

- Some extra assertions added to test build job.
- Manually tested some basic installation scenarios.

### Additional Notes

Note that the version provided as AGENT_VERSION and which is used as the
version for the "depends" field as well as for the version given to the
ddot-byoc package itself is arbitrary and holds no direct relationship
currently to the actual code checkout that the build is based on. But
this is consistent with what is currently done for the container BYOC
(where the agent image where the custom ddot is injected can be a
different version from the code used to build it). Therefore, I'm
leaving that for a possible follow up where this is "corrected" for both
containers and packages, to be discussed.

In addition, this currently doesn't properly support non-release
versions of the Agent (such as `devel`, `rc`s and the like). If we want
to support that, it would be a matter of sanitizing versions in the same
way as omnibus does.

[ABLD-102]:
https://datadoghq.atlassian.net/browse/ABLD-102?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ