[CWS-5084] Fix hardlinks dentry resolution (#40812)
### What does this PR do?

Hardlinks share the same inode and (potentially) the same mount ID.
This means that any hardlink path resolution will override any previous
path components of a different link for the same file, which makes the
resolution of hardlink paths racy and might lead to inconsistent path
resolutions.

For example executing the following: `/bin/cat /bin/gunzip` on busybox
(where both paths point to the same file) sometimes causes the exec
event path to be resolved to `/bin/gunzip` because the open syscall for
`/bin/gunzip` had already overwritten the dentry entries in the kernel
map at the time system-probe performs the dentry resolution of the exec
event.

To solve this, this PR introduces a new ID that is incremented every
time an inode with hardlinks is accessed.
This ID is merged in the existing path_id field to avoid increasing the
size of all kernel events.

Using a new ID avoids bumping the path_id map every time an inode with
hardlinks is accessed, meaning previous (non-hardlink) userspace cache
entries remain valid. This is at the downside of using less bits for the
path_id to account for dentry renaming and others.

### Motivation

### Describe how you validated your changes

### Additional Notes