[ACIX-923] Implement coverage computation for Kubernetes environment (#40791)
### What does this PR do?

Compute code coverage for Kubernetes based environment.
The coverage is computed on every agent pod and container in the cluster
and downloaded to local

Slightly update the interface to return an error and a string that
explain the result of the coverage collection. It should not fail if
part of the agent does not exist. Just print a summary to help
troubleshooting if needed.

### Motivation

Coverage should be used to dynamically determine relevant tests to
execute on PRs

### Describe how you validated your changes

### Additional Notes