[dyninst] Respect MaxCollectionSize and MaxLength (#40820)
https://datadoghq.atlassian.net/browse/DEBUG-4291

We used to propagate the collection capture limits statically during
irgen. With MaxCollectionSize and MaxLength limits, we need to be able
to have different limits between probes, while sharing the types with
the program. So now the irgen merely annotates types as dynamically
sized. Then the probes params carry the settings to the stack machine,
which uses the right setting for a given dynamic type when serializing
it.

I decided to preserve current default behavior, which is based on byte
size, not the element count. Maybe not worth the extra complexity
:shrug:.