[SBOM] make sure the overlayfs scanner uses the same lease during mount collection and scanning (#40911)
### What does this PR do?

- First commit:

Before this commit, the code was using one lease for the mount
collection, and another one for the scanning. Nothing guarantees that
the mounts collected in the first step are still valid during the second
step.

More importantly, the lease was always taken with the default
snapshotter, even though we also support the nydus one. Using the lease
from `Mounts` ensure we always use the correct one.

-  Second commit:

The cleanupTimeout was only used because of copy/paste from the mount
scan method. Let's also handle the case where the context doesn't have
a deadline.

### Motivation

### Describe how you validated your changes

### Additional Notes