[ AGNTLOG-303 ] Fix logs agent incompatibility with systemd >= 256 (#40747)
### What does this PR do?
A libzstd version mismatch within the agent code is causing issues with
newer versions of systemd, namely the
[datadog/zstd](https://github.com/DataDog/zstd) go dependency is set
[within the
agent](https://github.com/DataDog/datadog-agent/blob/1db7ec98d4f91cca55b21f1f616094ac9efeb723/go.mod#L54)
to v1.5.6 but the embedded libzstd.so.1 we're packaging with the agent
is set to v1.5.5. This can cause a segfault on all versions of systemd
>= 256, because [version
256](https://github.com/systemd/systemd/releases/tag/v256-rc1)
introduced a change to how the systemd library dynamically links to
local libzstd libraries.

In order to resolve this issue, this MR seeks to align the two versions
of libzstd together onto version 1.5.6

### Motivation

### Describe how you validated your changes
Agent versions deployed with this change onto journald-enabled systems
that reliably experience the segfault resolve the segfault.

Additionally, the expectation is that all libraries currently attempting
to load version 1.5.5 via dynamic linking are actually getting the
statically compiled 1.5.6 in its entirety. Any library attempting to
dynamically load 1.5.5 is expected to break in the same manner that
systemd does. Correspondingly, updating the 1.5.5 version to 1.5.6 is
locked to be either a fix or a no-op, rather than a potential avenue for
expected version mismatch.

### Additional Notes
resolves #40416

(cherry picked from commit 095ba886cb35fa5ed428759fc7b86312f0dd0c8f)