[CWS] implement SBOM resolver based on trivy-less collector (#40928)
### What does this PR do?

We currently ship trivy twice in the agent, once in the core agent
driving the SBOM related features, and once in the system-probe driving
some CWS features related to installed packages.

The use of trivy on the system-probe side is absolutely overkill and we
use only a very small subset of its features. This PR extracts the OS
package scanning to a V2 collector, that reads the files directly, and
respects an interface similar to the one currently provided by trivy.

This PR adds a config flag (disabled by default) to switch to using this
new collector. The SBOM usage in the system-probe is itself gated on a
config that is disabled by default as well, so the risk is very low
here.

This new collector already passes all the current functional tests in
the CWS testsuite, some follow up PRs will expand this with additional
tests.

### Motivation

### Describe how you validated your changes

### Additional Notes