add permission check on for ootb CR collection (#40894)
This PR adds a permission check to the out-of-the-box CR collection. If
the Cluster Agent doesn’t have the required permissions, it will skip
collecting that kind of custom resource.

The check is done by making a single List request to the api server and
checking if the returned error is a `forbidden` error. It doesn’t
require extra permissions and is simple to implement.

- If the cluster doesn’t support the specified custom resource kind in
the out-of-the-box collection, we log info
`Skipping built-in CR collector: no supported version found` once and
skip it.

- If the cluster doesn’t support the specified custom resource group in
the out-of-the-box collection, we log info
`Unsupported built-in CR collector` once and skip it.

- If the Cluster Agent doesn’t have permission to access a custom
resource in the out-of-the-box collection, we log info
`Skipping built-in collector due to insufficient permissions:` once and
skip it.

- If a user manually configures the Agent to collect a CR but it lacks
permissions, we log a warning every 15s:
`check:orchestrator | Collector argoproj.io/v1alpha1/rollouts is
skipped: couldn't sync informer argoproj.io/v1alpha1/rollouts in
1m10.001023574s`

Add environment var to cluster agent
```
clusterAgent:
  env:
    - name: DD_ORCHESTRATOR_EXPLORER_CUSTOM_RESOURCES_DATADOG_ENABLED
      value: true
    - name: DD_ORCHESTRATOR_EXPLORER_CUSTOM_RESOURCES_THIRD_PARTY_ENABLED
      value: true
```

Update cluster role to have incorrect role for argo rollouts
`rolloutsss`

```
- apiGroups:
  - datadoghq.com
  resources:
  - '*'
  verbs:
  - list
  - watch
- apiGroups:
  - karpenter.sh
  resources:
  - '*'
  verbs:
  - list
  - watch
- apiGroups:
  - karpenter.k8s.aws
  resources:
  - '*'
  verbs:
  - list
  - watch
- apiGroups:
  - karpenter.azure.com
  resources:
  - '*'
  verbs:
  - list
  - watch
- apiGroups:
  - argoproj.io
  resources:
  - rolloutsss
  verbs:
  - list
  - watch

```

Check cluster agent logs `Skipping built-in collector due to
insufficient permissions:`