[e2e] add additional error context when unmarshalling SBOMs fails (#40970)
### What does this PR do?

We recently started to see a bit of flakyness when unmarshalling SBOMs
in `TestSBOM` in new-e2e-containers. The error usually looks like
```
        	Error:      	Received unexpected error:
        	            	proto: cannot parse invalid wire-format data
        	Messages:   	Failed to query fake intake
```
the issue is that the agent sends directly the result of `proto.Marshal`
so this error is pretty suspicious. To help further investigation, this
PR adds additional context to the error: the encoding used, the size of
the payload, and a few bytes from the payload (just in case the payload
is reporting a encoding different than the one actually seen in the
data).

### Motivation

### Describe how you validated your changes

### Additional Notes