[CWS] make it possible to build system-probe without trivy (#40971)
### What does this PR do?

This PR extracts the trivy parts from the CWS SBOM resolver, making it
possible to be used without importing trivy at all. The trivy collector
is extracted into a different file, the build tags are removed where
possible, and the selection is done depending on the trivy availability
or not.

No functional change

### Motivation

### Describe how you validated your changes

### Additional Notes