dyninst/decode: fix error handling around slices (#40986)
### What does this PR do?

Before this change, we'd attempt to close the array and write that
something was pruned when we hit an error. This isn't safe because the
encoder will be in an undefined state.

Let's at least get a good error back out when this happens.

### Motivation

We saw this error in staging. I suspect something related to running out
of buffer space leads to this, but I haven't reproduced it. We need to
add more testing of running out of space in bpf and being robust to
those messages.