Default replacement value to $1 (#40649)
### What does this PR do?

This makes the `match_value` default to `"$1"` when `match_pattern` is
set, instead of defaulting to `""`.

### Motivation

Without this, the behavior of `match_pattern` without `match_value` is
to silently discard the entire value; I don't believe this is ever
something a customer would do intentionally.

With this change, setting `match_pattern` without `match_value` is
exactly the same as setting `extract_value` (and if we want, we could
deprecate `extract_value` entirely in the future).

### Describe how you validated your changes

I added a test for it and confirmed all the tests still pass. This
shouldn't change the behavior of the agent in any case except the
pathological one where someone has forgotten to set `match_value`.