[EBPF-808] gpu: use pools for spans and enriched kernel launches (#40445)
### What does this PR do?

Enables the use of pools for objects that are allocated in the heap in
the consumer hot path. These were causing a high number of allocations
during runtime, which caused GC pressure and high CPU usage.

To monitor leaks and the state of the pools, this PR also includes a
`TypedPool` structure with telemetry.

### Motivation

Reduce CPU usage in prod.

### Describe how you validated your changes

Included tests to ensure that objects are released back to the pool
properly in all cases.

Benchmarked in load-test environment, showing that for high workloads we
have a CPU usage reduction


![system-probe avg k8s CPU Usage by
env.png](https://app.graphite.dev/user-attachments/assets/57cb2ff2-ee52-4c4f-8433-6e74ecc8828c.png)

### Additional Notes