service discovery: Report raw DD_SERVICE name (#40815)
### What does this PR do?

There is a field set by the service discovery module which reports the
service name from the DD_SERVICE.  However, this field does not actually
correspond to the UST DD_SERVICE environment variable.

In the case of PHP and Java, it tries to get the server name from the
tracer configuration on the command line and reports that as the
DD_SERVICE-based name if it exists.  However, this is incomplete since
in both PHP and Java the properties that it checks for can be defined in
other places than the command line (like php.ini, web server config,
etc).  So it would be better to drop this logic and only report the name
from the environment variable as DD_SERVICE.

Also, it tries to read the extract the service from the DD_TAGS, which
is not a UST variable, but an environment used by tracing libraries.
This is also incomplete, since, for example, the Java or .NET tracers
can get the tags from the dd.tags system property instead, and that case
is not handled.

The service name from the tracing libraries can be obtained from the
tracer metadata which was introduced after this handling of DD_SERVICE.
That gets the name regardless of how the user configured it in the
tracing library.

Also, the code normalizes the name gotten from this sources but it would
be better to report the original name as specified by the user and allow
the backend to normalize it if it turns out to be necessary, so that the
agent doesn't drop information.  Note that currently normalization is
inconsistent between different products so such a decision is better
taken in the backend where it is more flexible.  For example, the
normalization in the agent code doesn't match the normalization done in
the logs backend.  For a service of foo😀😀ar, the code in the agent
results in foo_ar while the logs backend generates foo____ar.

Report the raw DD_SERVICE name from the UST environment variable only.
When the name comes from parsing the command line, use that as the
generated name.

### Motivation

https://datadoghq.atlassian.net/browse/DSCVR-183

### Describe how you validated your changes

Unit tests modified.

### Additional Notes