[dynamic test] Add a diffed index to dynamic test evaluation (#40849)
### What does this PR do?

Add a new index that is supposed that rely on the difference of coverage
between a baseline test that does nothing, and the actual coverage for a
test.
The underlying goal is to be less sensitive on PR that are updating part
of the code that is always executed, to avoid running all the tests all
the time in that case.

### Motivation

Run less and only relevant tests on PR

### Describe how you validated your changes

### Additional Notes