dyninst/procmon: add Sync support, replace ebpf/process
The abstraction was not adding much of anything, and was hurting in that we
were never getting exec calls that happen after the first one. Processes can
change their shape, and we need to re-analyze them.

The intention to avoid re-processing was fine, but it is misguided. We know
that we might miss exec events from eBPF, so we must re-process if it's
possible that the process changed. We could use some more memory to track
information about the fingerprint when we last did the analysis, but that
doesn't seem worthwhile.

Instead we make sure that we don't flood the queue from our syncing. On
the whole this should retain substantially less memory.

This also fixes the flakes we saw in the system tests.