process_log: Add process tags (#40800)
### What does this PR do?

Add any available process tags to the logs configurations generated when
we dynamically discover log files being written on the system.

Unlike in the case of container logs discovery, multiple processes could
be writing to the same log files.  We get the tags from the first such
process, which matches how we handle the existing service and source
fields in the log configuration.

Changes to other providers and generic code is needed to pass along the
tagger component to the process_log provider.

An alternative implementation could have been to have the tagger called
from pkg/logs/internal/tag with a process entity, however, that would be
more complex given that that there could be multiple processes (possibly
hundreds in the case of worker processes) per log file and that the
initial process could no longer exist.  In this chosen implementation,
we do not handle the case where different processes writing to the log
file have different tags.

### Motivation

https://datadoghq.atlassian.net/browse/DSCVR-191

### Describe how you validated your changes

Unit tests and E2E tests included.