dyninst/irgen: make sure all types get expanded
On a somewhat deep level, we need to visit all types we think should exist at least
with a depth of one to make sure that things like maps are well-formed. This reworks
the type exploration to make that unconditionally true. This fixes issues involving
failure to explore return types.