[dynamic-test] Support wildcards for target tests, and for package paths (#41101)
### What does this PR do?

Add support for wildcard in test targets, so that we can indicate that
all the tests should be executed on some modifications. For example on
framework update on on test definition change.

Also support specifying path with wildcard, so that we can indicate that
any modification to a file in `test/new-e2e/*` folder should trigger
some test.


Also update index creation to indicate that some path should trigger all
the test, depending on the targetted test.

### Motivation

Add some features to the index to suit more use case we have.

### Describe how you validated your changes

### Additional Notes