[corechecks/kubelet] Use workloadmeta as source for pods (#40563)
### What does this PR do?

This PR updates the Kubelet check to use workloadmeta instead of querying the Kubelet directly for pod info.
This only changes the pod provider of the check. The rest are already using workloadmeta or query the Kubelet to fetch info that's not pod related and it's not stored in workloadmeta.

This change is part of the effort to make workloadmeta the single source of truth for pod data. Benefits include:
- Consistent data across Agent components.
- Code easier to maintain and reason about.
- Improved performance by avoiding redundant data and extra calls to the Kubelet.

This is similar to the migration that was done for the AD prometheus provider (https://github.com/DataDog/datadog-agent/pull/39238) and the KSM check (https://github.com/DataDog/datadog-agent/pull/39657).

### Describe how you validated your changes

Relying mostly on existing tests plus manual checks to confirm the Kubelet check emits the same metrics.