[CWS] collectorv2/dpkg: ignore md5sums in status.d directory (#41184)
### What does this PR do?

On [distroless](https://github.com/GoogleContainerTools/distroless) images, all the status files and the md5sums files are in the `status.d` directory. This is handled correctly by the v2 collector when listing the installed files for packages, but we also currently try to read md5sums files when trying to read status files in `status.d`. This PR fixes this by skipping md5sums files when reading the status files.

### Motivation

### Describe how you validated your changes

### Additional Notes