smp: clean up image publishing to multiple regions
Rather than add a GitLab job for each region to which we would like to
push, this commit instead leverages the `IMG_REGISTRIES` parameter of
the `.docker_publish_job_definition` job and the `dd-image` command in
the `DataDog/public-images` repo. Since
<https://github.com/DataDog/public-images/pull/88> was merged, all of
the regions `smp` supports have the `smp` tag, so setting
`IMG_REGISTRIES` to `"smp"` will push Agent images to all of these
regions.

Thanks go to Lenaic Huard for the suggestion, which yields a much
cleaner change.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>