[CONTP-982] Expose services in verbose configcheck command (#41149)
### What does this PR do?

Prints out the list of active services identified by autodiscovery when using the `agent configcheck -v` command. get

### Motivation

Provides more insight into the state of agent autodiscovery in relation to to the checks that are (un)matched.

### Describe how you validated your changes

- Updated `TestPrintConfigCheck` test case.
- Manually QAd by running `agent configcheck -v`
<img width="1233" height="445" alt="image" src="https://github.com/user-attachments/assets/71dabbd3-35b5-4805-bb13-b91b2ab4ceba" />

### Additional Notes

This change also impacts flare output since `configcheck` is called with debug/verbose mode enabled when generating a flare.