[CONTP-988] Prioritize LocalData over ExternalData for Tag Enrichment (#41195)
### What does this PR do?

Prioritizes the use of LocalData over ExternalData if there is any mismatch between the containerIDs used to query for tags in the tagger.

### Motivation

Avoid adding improper tag information to tagger enriched telemetry. During high process/container turnover, there are a few cases in which external data may resolve an incorrect containerID which results in another workload's tags being added to telemetry. This can cause monitors and alarms to trigger when unexpected tag contexts are emitted.

### Describe how you validated your changes

CI and Unit Tests

### Additional Notes