disable virtual rendering to allow ctrl-f on full editor (#41148)
### What does this PR do?
Allows CTRL-F to work on the full settings page code editor. 

### Motivation
Easily search and find items.

### Describe how you validated your changes
Comparing CTRL-F functionality on a large configuration file before and after the change.

### Additional Notes
This PR essentially disables virtualization on the page and loads everything in to be able to search for it. The reason why it "wasn't" working prior was due to the search only works on rendered DOM items and CodeMirror only renders visible items by default. Setting the viewportMargin to infinity forces CodeMirror to load the entire configuration which may impact performance.