[trace-agent] Remove deprecated http client config (#41225)
### What does this PR do?
Work for https://datadoghq.atlassian.net/browse/APMSP-1715

Related datadog-exporter PR (merged): https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/42896

Remove HTTPClientFunc from trace-agent configuration. This setting was deprecated in favor of HTTPTransportFunc, which is now used in both the trace-agent and the OTel exporter. It is safe to remove it from the codebase.

### Motivation
See above

### Describe how you validated your changes
Existing tests. This setting is unused.

### Additional Notes